Added state management for VM instances
Add a check for chosen VM's state. This means that if a VM is already
created but in a state other than "Running" it will be started before
moving on to the next step.

Renamed fqdn to fullyQualifiedDomainName.
Renamed passwd to password.
Renamed hostn to hostname.

Added comments to multiple steps of the process.